techpack: audio: q6voice: Force AVCS CVP version 1 on legacy ADSP
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on all legacy ADSP firmwares: these ones cannot execute
the Q6 Service Version command on their Q6 DSP because of older
firmware not supporting it.
We know that for the old firmware the version is always 1, hence
forcing it is totally safe.

Signed-off-by: Jacob Deepu <d2jdeepu@gmail.com>
Signed-off-by: sweetyicecare <gabrielseedev@outlook.com>